+++ app/admin/course_events.rb
@@ -6,15 +6,24 @@ ActiveAdmin.register CourseEvent do
   #
   # Uncomment all parameters which should be permitted for assignment
   #
-  # permit_params :course_id, :room_id, :trainer_id, :course_schedule_id, :event_date
+  permit_params :category_id, :course_id, :room_id, :trainer_id, :course_schedule_id, :event_date, :state,
+                attendees_attributes: [ :id, :_destroy, :presence ]
   #
   # or
 
-  permit_params do
-    permitted = [:category_id, :course_id, :room_id, :trainer_id, :course_schedule_id, :event_date, :state]
-    permitted << :organization_id if current_admin_user.is_root? || params[:action] == 'create'
-    permitted
-  end
+  # permit_params do
+  #   permitted = [:category_id, :course_id, :room_id, :trainer_id, :course_schedule_id, :event_date, :state,
+  #                attendees_attributes: [ :id, :_destroy, :presence ]
+  #   ]
+  #   permitted << :organization_id if current_admin_user.is_root? || params[:action] == 'create'
+  #   permitted
+  # end
+
+  # before_update do |resource|
+  #   if resource.attendees.where(presence: true).count > 0
+  #     resource.auditor_id = current_admin_user.id # Not for admin user!
+  #   end
+  # end
 
   controller do
     def scoped_collection
@@ -70,6 +79,7 @@ ActiveAdmin.register CourseEvent do
           row(:course)
           row(:room)
           row(:trainer)
+          row(:auditor)
           row(:course_schedule)
           row(:event_date)
           row(:state) {|obj| span obj.localized_state, class: "status_tag #{obj.state}" }
@@ -80,7 +90,15 @@ ActiveAdmin.register CourseEvent do
       column do
         panel "#{Attendee.model_name.human(count: 2)} #{course_event.attendees.count} / #{course_event.room.max_attendees}" do
           table_for course_event.attendees do
+            column :id
             column :user
+            column :presence
+            column :created_at
+            column do |obj|
+              span link_to 'Mostra', [:admin, obj]
+              span link_to 'Modifica', edit_admin_attendee_path(obj)
+              # span link_to 'Elimina', destroy_admin_attendee_path(obj)
+            end
           end
         end
       end
@@ -88,21 +106,36 @@ ActiveAdmin.register CourseEvent do
   end
 
   form do |f|
-    f.inputs do
-      f.semantic_errors *f.object.errors.keys
-      if current_admin_user.is_root?
-        f.input :organization
-      else
-        f.input :organization, collection: [current_admin_user.organization]
+    columns do
+      column do
+        f.inputs do
+          f.semantic_errors *f.object.errors.keys
+          if current_admin_user.is_root?
+            f.input :organization
+          else
+            f.input :organization, collection: [current_admin_user.organization]
+          end
+          f.input :category, collection: current_admin_user.categories
+          f.input :course, collection: current_admin_user.courses
+          f.input :room, collection: current_admin_user.rooms
+          f.input :trainer, collection: current_admin_user.trainers
+          f.input :course_schedule, collection: f.object.course_id ? current_admin_user.course_schedules.where(course_id: f.object.course_id) : current_admin_user.course_schedules
+          f.input :event_date, as: :date_time_picker
+          f.input :state
+        end
       end
-      f.input :category, collection: current_admin_user.categories
-      f.input :course, collection: current_admin_user.courses
-      f.input :room, collection: current_admin_user.rooms
-      f.input :trainer, collection: current_admin_user.trainers
-      f.input :course_schedule, collection: f.object.course_id ? current_admin_user.course_schedules.where(course_id: f.object.course_id) : current_admin_user.course_schedules
-      f.input :event_date, as: :date_time_picker
-      f.input :state
+      column do
+        f.has_many :attendees, allow_destroy: can?(:destroy, f.object), new_record: can?(:create, f.object) do |ff|
+          ff.semantic_errors *ff.object.errors.keys
+          tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
+          ff.input :user_id, as: :search_select, url: admin_users_path(tmp_params),
+                  fields: [:email], display_name: 'email', minimum_input_length: 3,
+                  order_by: 'email_asc'
+          ff.input :presence
+        end
+      end
     end
+
     f.actions
   end
 end