+++ app/models/course_event.rb
@@ -7,8 +7,11 @@ class CourseEvent < ApplicationRecord
   belongs_to :room
   belongs_to :trainer
   belongs_to :course_schedule
+  belongs_to :auditor, class_name: 'User', optional: true
   has_many :attendees, dependent: :destroy
   has_many :users, through: :attendees
+
+  accepts_nested_attributes_for :attendees, reject_if: lambda { |obj| obj[:user_id].blank? }, allow_destroy: true
 
   validates :course_id, uniqueness: { scope: [:room_id, :trainer_id, :event_date],
                                  message: 'already exists' }